[TG-129] fix: webaddmin merge with develop and users

+++ src/main/java/com/safesteps/backend/domain/users/dto/AdminUserDTO.java
@@ -11,6 +11,6 @@ public class AdminUserDTO {
     private String pictureUrl;
     private Integer points;
     private Long level;
-    private Integer reputacio;
+    private double reputacio;
     private UserStatus status;
 }
